Fix OSError Exception
This commit fixes an issue where an OSError exception was raised when
trying to respond to the client after receiving data from the client.

This was caused by the use of a context manager in the receive function,
causing the socket to close prematurely, preventing further socket
methods from being used.

To fix this issue, the context manager in the receive function has been
removed.

Resolves #4.