Add Regex Matching
This commit adds a module to regex match expressions that the client
will send to the server.
As with any protocol, there must be an agreed way to communicate.

ADD_ENTRY ... will request the server to add an entry to the database.
GET_ENTRIES ... will request the server to respond with x entries.
GET_ENTRIES is hardwired to 10 as no other amount is needed.