Add Database Task Queuing
A queue has been implemented to make sure the database doesn't get
accessed by 2 or more threads at once, since only 1 connection is
allowed to the database at any 1 time.

To achieve this:
An additional thread which handles the database interactions is created.
Client handler threads will each have a unique identifier (ID).
After they receive the payload, they put it in the queue with its ID.
The database thread will get a task from the queue, match and run the
command (if valid) by calling the match function and save the match
function's results in the task_results dictionary with the identifier as
the key.
The client thread will call a function, get_results with its ID to fetch
the results from the dictionary and subsequently delete the dictionary
entry and release the ID. It responds with the response as usual and
closes the connection.

Only the database thread running the handle_db function calls the match
function which calls the database functions, this prevents multiple
connections to the database at 1 instant.

Resolves #7.